+++ src/app/project/[slug]/notice/page.tsx
@@ -1,23 +1,23 @@
 'use client';
 
 import { useRecoilValue } from 'recoil';
-import { projectNoticeActiveMenuStateStore } from '@/store/project/alert/AlertNavTabStateStore';
-import VAlertRecruitList from '@/components/project/alert/vote/recruit/list/VAlertRecruitList';
-import VAlertFWList from '@/components/project/alert/vote/fwithdraw/list/VAlertFWList';
-import VAlertFwModal from '@/components/project/alert/vote/fwithdraw/modal/VAlertFWModal';
-import VAlertRecruitModal from '@/components/project/alert/vote/recruit/modal/VAlertRecruitModal';
-import AlertCrewList from '@/components/project/alert/crew/list/AlertCrewList';
+import RCVoteNoticeList from '@/features/project/auth/myProject/notice/components/rcVoteNotice/RCVoteNoticeList';
+import VAlertFWList from '@/features/project/auth/myProject/notice/components/fwVoteNotice/FWVoteNoticeList';
+import FWVoteNoticeModal from '@/features/project/auth/myProject/notice/components/fwVoteNotice/FWVoteNoticeModal';
+import RCVoteNoticeModal from '@/features/project/auth/myProject/notice/components/rcVoteNotice/RCVoteNoticeModal';
+import CrewNoticeList from '@/features/project/auth/myProject/notice/components/crewNotice/CrewNoticeList';
+import { activeNoticeTabStateStore } from '@/features/project/auth/myProject/notice/store/ActiveNoticeTabStateStore';
 
 function NoticePage() {
-  const activeNoticeMenu = useRecoilValue(projectNoticeActiveMenuStateStore);
+  const activeNoticeMenu = useRecoilValue(activeNoticeTabStateStore);
 
   return (
     <section className='mb-20 tablet:basis-4/5'>
-      {activeNoticeMenu.name === '모집' && <VAlertRecruitList />}
+      {activeNoticeMenu.name === '모집' && <RCVoteNoticeList />}
       {activeNoticeMenu.name === '강제탈퇴' && <VAlertFWList />}
-      {activeNoticeMenu.name === '크루' && <AlertCrewList />}
-      <VAlertFwModal />
-      <VAlertRecruitModal />
+      {activeNoticeMenu.name === '크루' && <CrewNoticeList />}
+      <FWVoteNoticeModal />
+      <RCVoteNoticeModal />
     </section>
   );
 }
