--- src/app/project/_utils/constant.ts
@@ -1,6 +1,0 @@
-export const PROJECT_HISTORY_STATUS = {
-  FORCED_WDL: 'FORCED_WITHDRAWAL',
-  WDL: 'WITHDRAWAL',
-  PAR: 'PARTICIPATING',
-  FIN: 'FINISH',
-} as const;
